Rename electric gate, and add some minor refactoring. Also remove runtime scaling and ensure its edit mode only
